backend/requirements.txt
# Core dependencies
Django==3.0.5
djangorestframework==3.11.0
djoser==2.1.0
django-cors-headers==3.9.0
django-filter==21.1
django-extra-fields==3.0.2

# Database
psycopg2-binary==2.8.5

# Authentication
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0

# Image processing
Pillow==8.4.0

# PDF generation
reportlab==3.6.1

# ASGI/WSGI
gunicorn==20.1.0
asgiref==3.4.1

# Security
cryptography==3.4.8

# Utilities
requests==2.26.0
python3-openid==3.2.0
defusedxml==0.7.1
chardet==4.0.0
certifi==2021.5.30